One bug fix and Stage 2 of the file loading/spectrum processing optimization.
The bug fix is that Kojak now processes multiple crosslinkers. The functionality was lost somewhere along the way and is now back in place.

The loading/processing is fully multithreaded and mostly pipelined. I'm still unhappy in places. Also, optimizations were made to speed up the processing functions, though all testing was done in VS2013 (yikes!!) and might not translate to other compilers.

So much cruft was left behind. Next commit might be a code cleanup.